10/17/23
Purchased and imported the animations that will be used for the one handed sword mechanic.

Took me hours but I think I have finally worked out how to tie all of the animations in together. I created 3 different blend spaces for the animations for start loop and stop. I should be able to create links between them in the animation blue print